requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-generativeai==0.3.1
redis==5.0.1
rq==1.15.1
discord.py==2.3.2
kubernetes==28.1.0
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
